DRTII-849 Move shared arrival generator into its own file
