+++ server/src/test/scala/drt/shared/FlightDisplayFilterSpec.scala
@@ -1,9 +1,7 @@
 package drt.shared
 
-import controllers.ArrivalGenerator
 import drt.shared.CrunchApi.MillisSinceEpoch
-import drt.shared.Terminals.{T1, T2, T3, T4, T5, Terminal}
-import drt.shared.api.Arrival
+import drt.shared.Terminals._
 import org.specs2.mutable.Specification
 
 import scala.collection.immutable.HashSet
@@ -70,60 +68,4 @@ class FlightDisplayFilterSpec extends Specification {
       filter.forTerminal(flightsPostT4Opening, T5) === List(redListT5postT4, nonRedListT5postT4)
     }
   }
-}
-
-object ArrivalGenerator {
-  def arrival(
-               iata: String = "",
-               icao: String = "",
-               sch: MillisSinceEpoch = 0L,
-               actPax: Option[Int] = None,
-               maxPax: Option[Int] = None,
-               terminal: Terminal = Terminal("T1"),
-               origin: PortCode = PortCode(""),
-               operator: Option[Operator] = None,
-               status: ArrivalStatus = ArrivalStatus(""),
-               est: MillisSinceEpoch = 0L,
-               act: MillisSinceEpoch = 0L,
-               estChox: MillisSinceEpoch = 0L,
-               actChox: MillisSinceEpoch = 0L,
-               gate: Option[String] = None,
-               stand: Option[String] = None,
-               tranPax: Option[Int] = None,
-               runwayId: Option[String] = None,
-               baggageReclaimId: Option[String] = None,
-               airportId: PortCode = PortCode(""),
-               feedSources: Set[FeedSource] = Set(),
-               pcpTime: Option[MillisSinceEpoch] = None
-             ): Arrival =
-    Arrival(
-      Operator = operator,
-      Status = status,
-      Estimated = if (est != 0L) Some(est) else None,
-      Actual = if (act != 0L) Some(act) else None,
-      EstimatedChox = if (estChox != 0L) Some(estChox) else None,
-      ActualChox = if (actChox != 0L) Some(actChox) else None,
-      Gate = gate,
-      Stand = stand,
-      MaxPax = maxPax,
-      ActPax = actPax,
-      TranPax = tranPax,
-      RunwayID = runwayId,
-      BaggageReclaimId = baggageReclaimId,
-      AirportID = airportId,
-      Terminal = terminal,
-      rawICAO = icao,
-      rawIATA = iata,
-      Origin = origin,
-      PcpTime = if (pcpTime.isDefined) Option(pcpTime.get) else if (sch != 0L) Some(sch) else None,
-      Scheduled = sch,
-      FeedSources = feedSources
-    )
-
-  def flightWithSplitsForDayAndTerminal(date: SDateLike, terminal: Terminal = T1): ApiFlightWithSplits = ApiFlightWithSplits(
-    arrival(sch = date.millisSinceEpoch, terminal = terminal), Set(), Option(date.millisSinceEpoch)
-  )
-
-  def arrivalForDayAndTerminal(date: SDateLike, terminal: Terminal = T1): Arrival =
-    arrival(sch = date.millisSinceEpoch, terminal = terminal)
 }
